option to smooth over empty responseText, nsITimer
With our otrs installation at work i often get an empty response text in
xmlhttprequests. Since i experience the same with the iPhone otrs app i assume
it's not unique to this firefox extension :).
I'm not sure whether it's our installation or a general bug in the
iPhoneHandle package. Anyway, i added an option to ignore that outcome and keep
the prevoius displayed data. Otherwise it's replaced with an "?" and a tooltip
text saying the responsetext was empty.
One of the guidelines at developer.mozilla.org that use of setTimeout is flagged
at code review. One should use nsITimer instead. So i did just that.
Also i rearranged some controls in the options dialog.